# public/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Clients App"
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/clients": {
      "get": {
        "description": "Returns all clients from the system including providers",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "A list of clients."
          }
        }
      },
      "post": {
        "description": "Create a new client.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Create new Client ",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Client"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Created client object"
          },
          "400": {
            "description": "Bad input"
          }
        }
      }
    },
    "/clients/{id}" :{
      "get": {
        "description": "Returns a client from the system including providers based on id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Client id ",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Client object."
          },
          "400" : {
            "description" : "Client not Found"
          }
        }
      },
      "put": {
        "description": "Edit a client.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Client id ",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Edit a Client ",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Client"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated client object"
          },
          "400": {
            "description": "Bad input"
          }
        }
      },
      "delete" :{
        "description": "Delete a client from the system based on id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Client id ",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Client object."
          },
          "400" : {
            "description" : "Client not Found"
          }
        }
      }
    },
    "/providers": {
      "get": {
        "description": "Returns all providers from the system",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "A list of providers."
          }
        }
      },
      "post": {
        "description": "Create a new providers.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Create new Client ",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Provider"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Created provider object"
          },
          "400": {
            "description": "Bad input"
          }
        }
      }
    },
    "/provider/{id}" :{
      "get": {
        "description": "Returns a provider from the system based on id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Provider id ",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Provider object."
          },
          "400" : {
            "description" : "Provider not Found"
          }
        }
      },
      "put": {
        "description": "Edit a provider.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Client id ",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Edit a Client ",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Client"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated client object"
          },
          "400": {
            "description": "Bad input"
          }
        }
      },
      "delete" :{
        "description": "Delete a client from the system based on id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Client id ",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Client object."
          },
          "400" : {
            "description" : "Client not Found"
          }
        }
      }
    }
  },
  "definitions" :{
    "ProviderId": {
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "example" : 1
        }
      }
    },
    "Provider": {
      "required": [
        "name"
      ],
      "properties": {
        "name": {
          "type": "string",
          "example": "Provider 1"
        }
      }
    },
    "Client": {
      "required": [
        "name",
        "phone",
        "email"
      ],
      "properties": {
        "name": {
          "type": "string",
          "example": "Client1"
        },
        "phone": {
          "type": "string",
          "example" : "12512512"
        },
        "email": {
          "type": "email",
          "example" : "client@email.com"
        },
        "Providers" : {
          "type" : "array",
          "items" : {
            "$ref": "#/definitions/ProviderId"
          }
        }
      }
    }
  }
}